% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulder.R
\docType{data}
\name{boulder}
\alias{boulder}
\title{City of Boulder, CO Waste Water Treatment}
\format{
An object of class \code{"list"} with seven data frames of flow data, phosphorous measurement, and grab sample results.
}
\source{
City of Boulder, CO
}
\usage{
data(boulder)
}
\description{
Flow data from Boulder Waste Water Treatment Plant with the goal of
predicting phosphorous in the effluent flow.
}
\keyword{datasets}
