% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulder_ammonia.R
\docType{data}
\name{boulder_ammonia}
\alias{boulder_ammonia}
\title{LIFT Wastewater Treatment Data}
\format{
boulder_ammonia is an object of class \code{"data.frame"} with 25908
rows and 16 columns.
}
\usage{
boulder_ammonia
data(boulder_ammonia) #- loads into working environment
}
\description{
The LIFT (Leaders Innovation Forum for Technology) data set
gives water quality measurements at numerous locations
in a water treatment facility. The facility is divided into
nine zones. Multiple probes are used measure dissolved oxygen,
ammonia, and nitrate.
}
\examples{
ammonia <- boulder_ammonia[,11]
length(ammonia)
ts.plot(ammonia)
ts.plot(ammonia[24000:25908])
hist(ammonia)
hist(ammonia,breaks="FD")
}
\keyword{datasets}
