% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denver_filters.R
\docType{data}
\name{denver_filters}
\alias{denver_filters}
\title{City of Denver, CO Filter Experiment}
\format{
An object of class \code{"list"} with sixteen filter data sets
and a backwash data set. Each data set has 105384 rows and variables including
Date.time, head loss, flow rate, turbidity, and filter run length
}
\source{
City of Denver, CO
}
\usage{
data(denver_filters)
}
\description{
One year of filter data for 16 different filters measured at 5 minute intervals
with the goal of comparing filters between each other and before/after
media replacement.
}
\keyword{datasets}
