% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_baseflow.R
\docType{data}
\name{DOC_baseflow}
\alias{DOC_baseflow}
\title{DOC Baseflow}
\format{
An object of class \code{"data.frame"} with 1174 rows and 32 variables including date, locations, Dissolved Organic Carbon, multiple environmental factors.
}
\source{
Unknown
}
\usage{
data(DOC_baseflow)
}
\description{
Time Series/Spatial dataset related to tracking Mountain Pine Beetles
}
\examples{
plot(DOC_baseflow$Latitude, DOC_baseflow$Longitude)
}
\keyword{datasets}
