% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_usgs.R
\docType{data}
\name{DOC_usgs}
\alias{DOC_usgs}
\title{DOC_usgs}
\format{
An object of class \code{"data.frame"} with 2900 rows and 6 variables including date, station location, DOC.
}
\source{
Unknown
}
\usage{
data(DOC_usgs)
}
\description{
Dissolved Organic Carbon spatial time series, station data
}
\examples{
plot(DOC_usgs$Latitude, DOC_usgs$Longitude)
}
\keyword{datasets}
