% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EML_day.R
\docType{data}
\name{EML_day}
\alias{EML_day}
\title{EML_day}
\format{
An object of class \code{"data.frame"} with 240 rows and 10 variables including date.time, depth, water quality measurements, and a depth group factor.
}
\source{
Professor Thad Scott from the Department of Biology at Baylor University provided this data.
Tarrant Regional Water District (TRWD) provided funding for the first year of data collection, and Eagle Mountain Lake is owned and operated by TRWD.
This lake is used to supply water to municipalities within the Dallas/Fort Worth area.
}
\usage{
data(EML_day)
}
\description{
Single day measurement of EML data on 4/25/2019
}
\examples{
EML_day$DepthLevels
}
\keyword{datasets}
