% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EML_week.R
\docType{data}
\name{EML_week}
\alias{EML_week}
\title{EML_week}
\format{
An object of class \code{"data.frame"} with 1764 rows and 11 variables including date.time, depth, water quality measurements, group factors.
}
\source{
Professor Thad Scott from the Department of Biology at Baylor University provided this data.
Tarrant Regional Water District (TRWD) provided funding for the first year of data collection, and Eagle Mountain Lake is owned and operated by TRWD.
This lake is used to supply water to municipalities within the Dallas/Fort Worth area.
}
\usage{
data(EML_week)
}
\description{
Single week of EML data with added group variables
}
\examples{
EML_week$TempNew
}
\keyword{datasets}
