% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\docType{data}
\name{freshwater}
\alias{freshwater}
\title{Freshwater}
\format{
An object of class \code{"data.frame"} with 662 rows and 5 variables including country, year, code, and withdrawal amount.
}
\usage{
data(freshwater)
}
\description{
Global water stress data, where freshwater withdrawal is measured as share of resources.
}
\keyword{datasets}
