% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIFT.R
\docType{data}
\name{LIFT}
\alias{LIFT}
\title{LIFT}
\format{
An object of class \code{"data.frame"} with 25908 rows and 15 variables measuring ammonia at different locations
}
\usage{
data(LIFT)
}
\description{
Water treatment facility data.
}
\keyword{datasets}
