% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines_fault.R
\docType{data}
\name{mines_fault}
\alias{mines_fault}
\title{Colorado School of Mines Waste Water Treatment Plant}
\format{
An object of class \code{"data.frame"} with 
43552 rows and 42 variables including
date.time, locations with water quality measurements, 
and true/false factors.
\describe{
  \item{Cyclic Variables}{these variables can be used 
  to model the daily and hourly trends in the data, 
  if such trends exist}
  \item{Control Variables}{these are variables whose 
  values are set by the operator: for example, 
  The blower flow rates, the transmembrane pressure, or 
  the “mode” of operation}
  \item{Response Variables}{these are variables that you 
  may wish to monitor after controlling for the effect 
  of the cyclic and control variables - examples are
  dissolved oxygen, level in tank, temperature, total 
  suspended solids in bioreactor, level in membrane tank,
  DO, pH, conductivity, TSS in the RAS (return activated 
  sludge, turbidity, and total suspended solids in permeate}
}
}
\usage{
mines_fault
data(mines_fault) #- loads into working environment
}
\description{
The mines fault data set comes from a water quality study from 
Mines Waste Water Treatment Plant, where the goal was to predict faults.
The original data are 43,552 measurements of 42 variables 
with observations are recorded every 10 minutes from April 10,
2010 at 6:10 AM to May 11, 2010 at 5:00 AM. 
A fault was identified at April 24, 2010, 10:00 AM. 
It took over two months for the system to fully recover.
}
\examples{
# fault on April 24, 2010, 10:00 AM.
month.val <- lubridate::month(mines_fault$Date.Time)
day.val   <- lubridate::day(mines_fault$Date.Time)
hour.val  <- lubridate::hour(mines_fault$Date.Time)
fault.time <- which(month.val==4 & day.val==24 & hour.val==10)[1]

# plot pH
# The biological community cannot survive if
# the pH is less than 6.5
# It clearly dips below this value
ts.plot(mines_fault$ras_ph, xlab="Index", ylab="pH")
abline(h=6.5,col=4,lwd=2)
abline(v=fault.time,col=2,lwd=2)

# plot MBR flow
ts.plot(mines_fault$mbr_1_perm_flow, xlab="Index", ylab="MBR 1, Permeate Flow")
abline(v=fault.time,col=2,lwd=2)

ts.plot(mines_fault$mbr_2_perm_flow, xlab="Index", ylab="MBR 2, Permeate Flow")
abline(v=fault.time,col=2,lwd=2)

# sort variables into groups
# cyclic variables
cyclic_vars <- c("cos_daily", 
                 "sin_daily", 
                 "cos_2hour", 
                 "sin_2hour", 
                 "cos_hourly", 
                 "sin_hourly")
                 
# control variables
control_vars <- c("bio_1_blow_flow", 
                  "bio_2_blow_flow", 
                  "mbr_1_tmp", 
                  "mbr_2_tmp",
                  "ambient_temp", 
                  "bio_1_phase_1", 
                  "bio_1_phase_2", 
                  "bio_2_phase_1",
                  "bio_2_phase_2", 
                  "mbr_1_mode_1", 
                  "mbr_1_mode_2", 
                  "mbr_1_mode_4",
                  "mbr_2_mode_1", 
                  "mbr_2_mode_2", 
                  "mbr_2_mode_4")

# response variables
response_vars <- c("mbr_1_perm_flow", 
                   "mbr_2_perm_flow", 
                   "ras_temp",
                   "bio_1_do", 
                   "bio_2_do", 
                   "mbr_1_level", 
                   "mbr_2_level", 
                   "perm_turb", 
                   "sewage_flow", 
                   "bio_1_level", 
                   "bio_2_level",
                   "bio_1_temp", 
                   "bio_2_temp", 
                   "bio_1_tss", 
                   "bio_2_tss", 
                   "perm_tank_level", 
                   "ras_do", 
                   "ras_ph", 
                   "perm_cond", 
                   "ras_tss")

# scale predictors                  
scale_predictors <- c("bio_1_blow_flow",
                      "bio_2_blow_flow", 
                      "ambient_temp",
                      "mbr_1_tmp", 
                      "mbr_2_tmp")
}
\keyword{datasets}
