% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesfault.R
\docType{data}
\name{minesfault}
\alias{minesfault}
\title{Colorado School of Mines Waste Water Treatment Plant}
\format{
An object of class \code{"data.frame"} with 43552 rows and 42 variables including
date.time, locations with water quality measurements, and true/false factors.
}
\usage{
data(minesfault)
}
\description{
Water quality dataset from Mines Waste Water Treatment Plant, used to predict faults.
}
\keyword{datasets}
