
\docType{package}
\name{mowateR}
\alias{mowateR}
\title{MoWaTER}
\description{
Data Sets from the Modernizing Water and Wastewater Treatment through Data Science Education and Research Team.
}
\details{
This package contains multiple datasets related to water and wastewater
treatment.
\describe{
\item{aqua_backwash}{}
\item{aqua_filter}{}
\item{bins}{}
\item{boulder_ammonia}{}
\item{boulder_phos}{}
\item{colorado_springs_phos}{}
\item{denver_filters (On Google Drive)}{}
\item{dissolved_carbon}{}
\item{eml}{}
\item{geosmin}{}
\item{hazen}{}
\item{lake_mead}{}
\item{mines_fault}{}
\item{orange_county}{}
\item{precip_type (On Google Drive)}{}
\item{reverse_osmosis}{}
\item{superior_aeration}{}
\item{tds_rivers}{}
}
}
