% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\docType{data}
\name{precip}
\alias{precip}
\title{Precipitation Study: precip}
\format{
An object of class \code{"data.frame"} with 7139951 rows and 5 variables:
prec type, date, station, altitude, temperature.
}
\usage{
data(precip)
}
\description{
Reshaped dataset from predictors, where each profile is split into 31 observations.
}
\keyword{datasets}
