% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_type.R
\docType{data}
\name{precip_type}
\alias{precip_type}
\title{Precipitation Study: predictors}
\format{
An object of class \code{"list"} with two \code{"lists"}:
\describe{
  \item{precip}{Station location and temperature data.}
  \item{predictors}{Temperature profiles, dates, locations, and
  precipitation type.}
  \describe{
    \item{Twb.prof}{the corresponding vertical profiles of
    temperature (0m to 3000m above the surface in steps of 100m,
    in Kelvin)}
    \item{date.ind}{for each case, the index of the date on
    which it was reported}
    \item{ptype}{cases where one of the four precipitation
    types of interest was reported}
    \item{station.ind}{for each case, the index of the station
    where it was reported}
    \item{stations}{names of all stations for which data are available}
  }
}
}
\source{
Michael Scheuerer from NOAA’s Earth System
Research Laboratories (ESRL) division, The US National
Atmospheric and Oceanic Administration (NOAA) https://www.ncdc.noaa.gov
}
\usage{
data(precip_type)
}
\description{
The precipitation type data set is a collection of data on weather
conditions across the continental United States and Alaska.
Precipitation type is important because certain types of
even moderate amounts can have a dramatic impact on public safety.
There are 18 years of data. At weather stations, one of four
precipitation types is observed: Rain, Snow, Ice Pellets, or
Freezing Rain.  At each weather station where a precipitation was
observed, the corresponding grid box (in space and time) of a numerical
weather prediction model is obtained. Then, the vertical temperature
profile from that grid box is extracted. The shape and position (relative
to the freezing temperature) of the vertical temperature profile
generally determines the type of precipitation that is observed.
}
\examples{
# profiles for Waco airport

which(precip_type$predictors$stations == "KACT    ")
waco_index <- which(precip_type$predictors$station.ind == 26)
waco_profiles <- precip_type$predictors$Twb.prof[waco_index,]
cols <- viridis::viridis(length(waco_index))
plot(waco_profiles[1,] ,1:31, type = "l",
     xlim = c(260, 300), col = cols[1],
     xlab = "", ylab = "")
for (i in 2:length(waco_index)) {
  lines(waco_profiles[i,], 1:31, col = cols[i])
}
}
\keyword{datasets}
