% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\docType{data}
\name{predictors}
\alias{predictors}
\title{Precipitation Study: predictors}
\format{
An object of class \code{"list"} with nine objects.
}
\usage{
data(predictors)
}
\description{
Time series/Spatial profile dataset using atmospheric temperature to predict precipitation type.
}
\keyword{datasets}
