% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROenergy.R
\docType{data}
\name{ROenergy}
\alias{ROenergy}
\title{ROenergy}
\format{
An object of class \code{"data.frame"} with 25844 rows
and 24 variables including date.time, runtime,
water quality measurements, and group factors.
}
\source{
Professor Tzahi Cath from the Department of Civil and Environmental Engineering at Colorado School of Mines.
}
\usage{
data(ROenergy)
}
\description{
Data for investigating reverse osmosis filter performance
}
\keyword{datasets}
