% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statewide.R
\docType{data}
\name{statewide}
\alias{statewide}
\title{statewide}
\format{
An object of class \code{"data.frame"} with around 315,000 obs. and 5 variables including date, storage amounts, capacity, percent full.
}
\usage{
data(statewide)
}
\description{
Estimates of Texas reservior data measured daily for 90 years
}
\keyword{datasets}
