% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Precipitation Study: stations}
\format{
An object of class \code{"data.frame"} with  551 rows and 4 variables: station, lon, lat, and elevation.
}
\usage{
data(stations)
}
\description{
Reshaped dataset from predictors, only considering locations.
}
\keyword{datasets}
