% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\docType{data}
\name{tds}
\alias{tds}
\title{Colorado River Total Dissolved Solids}
\format{
An object of class \code{"data.frame"} with 672 rows and 4 variables: year, date, tds in Lees and Hoover dams.
}
\source{
Unknown
}
\usage{
data(tds)
}
\description{
Time series of total dissolved solids in the Colorado river, measured at Lees and Hoover dams.
}
\examples{
plot(tds$year, tds$lees)
}
\keyword{datasets}
