% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_reservoirs.R
\docType{data}
\name{tx_reservoirs}
\alias{tx_reservoirs}
\title{statewide}
\format{
An object of class \code{"data.frame"} with around 315,000 obs. and 5 variables including date, storage amounts, capacity, percent full.
}
\usage{
data(tx_reservoirs)
}
\description{
Estimates of Texas reservoir data measured daily for 90 years
}
\keyword{datasets}
