% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbanrural.R
\docType{data}
\name{urbanrural}
\alias{urbanrural}
\title{Urban vs. Rural Improved Water Access Data}
\format{
An object of class \code{"data.frame"} with 5164 rows and
7 columns:
\describe{
\item{entity}{}
\item{code}{}
\item{year}{}
\item{urban}{Percent of urban population with access to clean drinking water}
\item{rural}{Percent of rural population with access to clean drinking water}
\item{pop}{Entity population}
}
}
\usage{
urbanrural
data(urbanrural) # loads into working evironment
}
\description{
The Urban vs. Rural data is a Time Series measuring water
availability in countries around the world. This data set contains
information on 205 entities from 1990 to 2015.
}
\keyword{datasets}
