% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\docType{data}
\name{bins}
\alias{bins}
\title{Bureau of Recreation Vegetation Bin Filter Experiment}
\format{
An object of class \code{"list"} with datasets containing flow and field data.
}
\source{
Colorado Bureau of Recreation
}
\usage{
data(bins)
}
\description{
Data from experimenting with using vegetation bins as filters.
}
\keyword{datasets}
