% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloth_filter.R
\docType{data}
\name{cloth_filter}
\alias{cloth_filter}
\title{Aqua Aerobic Cloth Filters}
\format{
An object of class \code{"list"} containing data sets of grab samples and high-density online data.
}
\source{
Aqua Aerobic
}
\usage{
data(cloth_filter)
}
\description{
Primary filtration data used to investigate cloth filter efficacy.
}
\keyword{datasets}
