% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolved_carbon.R
\docType{data}
\name{dissolved_carbon}
\alias{dissolved_carbon}
\title{DOC Baseflow}
\format{
An object of class \code{"data.frame"} with 1174 rows and 32 variables including date, locations, Dissolved Organic Carbon, multiple environmental factors.
}
\source{
Unknown
}
\usage{
data(dissolved_carbon)
}
\description{
Time Series/Spatial dataset related to tracking Mountain Pine Beetles
Dissolved Organic Carbon spatial time series, station data
}
\examples{
plot(dissolved_carbon[[2]]$Latitude, dissolved_carbon[[2]]$Longitude)
}
\keyword{datasets}
