% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\docType{data}
\name{eml}
\alias{eml}
\title{Eagle Mountain Lake Profile Data Sets}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(eml)
}
\description{
Data from the Eagle Mountain Lake
}
\examples{
data(eml)
}
\keyword{datasets}
