% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\docType{data}
\name{filters}
\alias{filters}
\title{filters}
\format{
An object of class \code{"data.frame"} with 35532 rows and 7 variables:
\describe{
\item{Date-Time}{Date and two-hour time period measurement was taken}
\item{Depth}{Profile Depth, measured in meters}
\item{Temp}{Water Temperature, measured in degrees Celsius}
\item{DO}{Dissolved Oxygen, measured in mg/L}
\item{DOsat}{Dissolved Oxygen Saturation, measured as the percentage of DO relative to what the concentration would be in equilibrium with the atmosphere, calcualted as actual/expected}
\item{pH}{pH, measured as standard pH from 0 (basic) to 14 (acidic)}
\item{Cond}{Conductivity in water, measured in \eqn{\mu}S/cm (micro Siemens per centimeter)}
}
}
\source{
Professor Thad Scott from the Department of Biology at Baylor University provided this data.
Tarrant Regional Water District (TRWD) provided funding for the first year of data collection, and Eagle Mountain Lake is owned and operated by TRWD.
This lake is used to supply water to municipalities within the Dallas/Fort Worth area.
}
\usage{
data(urbanrural)
}
\description{
Describe here
}
\keyword{datasets}
