% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIFT.R
\docType{data}
\name{LIFT}
\alias{LIFT}
\title{LIFT Wastewater Treatment Data}
\format{
LIFT is an object of class \code{"data.frame"} with 25908
rows and 16 columns.
}
\usage{
LIFT
data(LIFT) # loads into working environment
}
\description{
The LIFT (Leaders Innovation Forum for Technology) data set
gives water quality measurements at numerous locations
in a water treatment facility. The facility is divided into
nine zones. Multiple probes are used measure dissolved oxygen,
ammonia, and nitrate.
}
\examples{
ammonia <- LIFT[,11]
length(ammonia)
ts.plot(ammonia)
ts.plot(ammonia[24000:25908])
hist(ammonia)
hist(ammonia,breaks="FD")
}
\keyword{datasets}
