% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines_fault.R
\docType{data}
\name{mines_fault}
\alias{mines_fault}
\title{Colorado School of Mines Waste Water Treatment Plant}
\format{
An object of class \code{"data.frame"} with 43552 rows and 42 variables including
date.time, locations with water quality measurements, and true/false factors.
}
\usage{
data(mines_fault)
}
\description{
Water quality dataset from Mines Waste Water Treatment Plant, used to predict faults.
}
\keyword{datasets}
