% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mowater-mowater.R
\docType{package}
\name{mowater}
\alias{mowater}
\title{MoWaTER}
\description{
Data Sets from the Modernizing Water and Wastewater Treatment
through Data Science Education and Research Team.
}
\details{
This package contains multiple datasets related to water and wastewater
treatment.
\describe{
\item{bins}{Do we want to include mini-descriptions here?}
\item{boulder_ammonia}{}
\item{boulder_phos}{}
\item{cloth_filter}{}
\item{denver_filters}{}
\item{dissolved_carbon}{}
\item{eml}{}
\item{mines_fault}{}
\item{precip_type}{}
\item{reverse_osmosis}{}
\item{superior_aeration}{}
\item{tds_rivers}{}
\item{tx_reservoirs}{}
\item{urban_rural_water}{}
\item{water_stress}{}
}
}
