% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_type.R
\docType{data}
\name{precip_type}
\alias{precip_type}
\title{Precipitation Study: predictors}
\format{
An object of class \code{"list"} with nine objects.
}
\usage{
data(precip_type)
}
\description{
Time series/Spatial profile dataset using atmospheric temperature to predict precipitation type.
}
\keyword{datasets}
