% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_osmosis.R
\docType{data}
\name{reverse_osmosis}
\alias{reverse_osmosis}
\title{reverse osmosis}
\format{
An object of class \code{"data.frame"} with 25844 rows
and 24 variables including date.time, runtime,
water quality measurements, and group factors.
}
\source{
Professor Tzahi Cath from the Department of Civil and Environmental Engineering at Colorado School of Mines.
}
\usage{
data(reverse_osmosis)
}
\description{
Data for investigating reverse osmosis filter performance
}
\keyword{datasets}
