% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superior_aeration.R
\docType{data}
\name{superior_aeration}
\alias{superior_aeration}
\title{superior}
\format{
An objects of class \code{"list"} with two data frames totalling 750 rows and numerous date, water quality measurements, flow measurements, and group factor variables.
}
\source{
City of Superior Waste Water Treatment Plant.
}
\usage{
data(superior_aeration)
}
\description{
Flow data from Superior Waste Water Treatment Plant with the goal of
comparing outcome of two aeration basins. This data list is comprised
of two data sets.
}
\keyword{datasets}
