% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_rivers.R
\docType{data}
\name{tds_rivers}
\alias{tds_rivers}
\title{Colorado River Total Dissolved Solids}
\format{
An object of class \code{"data.frame"} with 672 rows
and 4 variables: year, date, tds in Lees and Hoover dams.
}
\source{
Unknown
}
\usage{
tds_rivers
data(tds_rivers) #- loads into working environment
}
\description{
Time series of total dissolved solids in the Colorado river, measured at Lees and Hoover dams.
}
\examples{
plot(tds_rivers$Date.Time, tds_rivers$lees, type = "l")
}
\keyword{datasets}
