% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbanrural.R
\docType{data}
\name{urbanrural}
\alias{urbanrural}
\title{urbanrural}
\format{
An object of class \code{"data.frame"} with 5164 rows and 7 variables including country and code, year (unbalanced, about 15 years per country), water quality for urban and rural areas, population.
}
\usage{
data(urbanrural)
}
\description{
Time series measuring water availability around the world.
}
\keyword{datasets}
